+++ angular/src/directives/navigation/ion-back-button.ts
@@ -1,4 +1,4 @@
-import { Directive, HostListener, Optional } from '@angular/core';
+import { Directive, HostListener, Input, Optional } from '@angular/core';
 import { AnimationBuilder } from '@ionic/core/components';
 
 import { Config } from '../../providers/config';
@@ -8,11 +8,12 @@ import { IonRouterOutlet } from './ion-router-outlet';
 
 @Directive({
   selector: 'ion-back-button',
-  inputs: ['defaultHref', 'routerAnimation'],
 })
-export class IonBackButtonDelegate {
-
+export class IonBackButtonDelegateDirective {
+  @Input()
   defaultHref: string | undefined | null;
+
+  @Input()
   routerAnimation?: AnimationBuilder;
 
   constructor(
@@ -25,10 +26,10 @@ export class IonBackButtonDelegate {
    * @internal
    */
   @HostListener('click', ['$event'])
-  onClick(ev: Event) {
+  onClick(ev: Event): void {
     const defaultHref = this.defaultHref || this.config.get('backButtonDefaultHref');
 
-    if (this.routerOutlet && this.routerOutlet.canGoBack()) {
+    if (this.routerOutlet?.canGoBack()) {
       this.navCtrl.setDirection('back', undefined, undefined, this.routerAnimation);
       this.routerOutlet.pop();
       ev.preventDefault();