+++ angular/src/directives/navigation/router-link-delegate.ts
@@ -1,5 +1,5 @@
 import { LocationStrategy } from '@angular/common';
-import { Directive, ElementRef, HostListener, Optional } from '@angular/core';
+import { ElementRef, OnChanges, OnDestroy, OnInit, Directive, HostListener, Input, Optional } from '@angular/core';
 import { Router, RouterLink } from '@angular/router';
 import { AnimationBuilder, RouterDirection } from '@ionic/core/components';
 import { Subscription } from 'rxjs';
@@ -8,13 +8,14 @@ import { NavController } from '../../providers/nav-controller';
 
 @Directive({
   selector: '[routerLink]',
-  inputs: ['routerDirection', 'routerAnimation']
 })
-export class RouterLinkDelegate {
-
+export class RouterLinkDelegateDirective implements OnInit, OnChanges, OnDestroy {
   private subscription?: Subscription;
 
+  @Input()
   routerDirection: RouterDirection = 'forward';
+
+  @Input()
   routerAnimation?: AnimationBuilder;
 
   constructor(
@@ -22,18 +23,18 @@ export class RouterLinkDelegate {
     private navCtrl: NavController,
     private elementRef: ElementRef,
     private router: Router,
-    @Optional() private routerLink?: RouterLink,
-  ) { }
+    @Optional() private routerLink?: RouterLink
+  ) {}
 
-  ngOnInit() {
+  ngOnInit(): void {
     this.updateTargetUrlAndHref();
   }
 
-  ngOnChanges(): any {
+  ngOnChanges(): void {
     this.updateTargetUrlAndHref();
   }
 
-  ngOnDestroy(): any {
+  ngOnDestroy(): void {
     if (this.subscription) {
       this.subscription.unsubscribe();
     }
@@ -50,7 +51,7 @@ export class RouterLinkDelegate {
    * @internal
    */
   @HostListener('click', ['$event'])
-  onClick(ev: UIEvent) {
+  onClick(ev: UIEvent): void {
     this.navCtrl.setDirection(this.routerDirection, undefined, undefined, this.routerAnimation);
     ev.preventDefault();
   }