+++ angular/src/directives/overlays/modal.ts
@@ -1,20 +1,68 @@
 /* eslint-disable */
 /* tslint:disable */
-import { ChangeDetectionStrategy, ChangeDetectorRef, Component, ContentChild, ElementRef, EventEmitter, NgZone, TemplateRef } from "@angular/core";
-import { ProxyCmp, proxyOutputs } from "../angular-component-lib/utils";
+import {
+  ChangeDetectionStrategy,
+  ChangeDetectorRef,
+  Component,
+  ContentChild,
+  ElementRef,
+  EventEmitter,
+  NgZone,
+  TemplateRef,
+} from '@angular/core';
+import { ProxyCmp, proxyOutputs } from '../angular-component-lib/utils';
 import { IonModal as IonModalCmp } from '@ionic/core/components/ion-modal.js';
 
 export declare interface IonModal extends IonModalCmp {}
 @ProxyCmp({
   tagName: 'ion-modal',
   customElement: IonModalCmp,
-  inputs: ["animated", "backdropBreakpoint", "backdropDismiss", "breakpoints", "cssClass", "enterAnimation", "event", "handle", "initialBreakpoint", "isOpen", "keyboardClose", "leaveAnimation", "mode", "presentingElement", "showBackdrop", "swipeToClose", "translucent", "trigger"], "methods": ["present", "dismiss", "onDidDismiss", "onWillDismiss"]
+  inputs: [
+    'animated',
+    'backdropBreakpoint',
+    'backdropDismiss',
+    'breakpoints',
+    'cssClass',
+    'enterAnimation',
+    'event',
+    'handle',
+    'initialBreakpoint',
+    'isOpen',
+    'keyboardClose',
+    'leaveAnimation',
+    'mode',
+    'presentingElement',
+    'showBackdrop',
+    'swipeToClose',
+    'translucent',
+    'trigger',
+  ],
+  methods: ['present', 'dismiss', 'onDidDismiss', 'onWillDismiss'],
 })
 @Component({
-  selector: "ion-modal",
+  selector: 'ion-modal',
   changeDetection: ChangeDetectionStrategy.OnPush,
   template: `<ng-container [ngTemplateOutlet]="template" *ngIf="isCmpOpen"></ng-container>`,
-  inputs: ["animated", "backdropBreakpoint", "backdropDismiss", "breakpoints", "cssClass", "enterAnimation", "event", "handle", "initialBreakpoint", "isOpen", "keyboardClose", "leaveAnimation", "mode", "presentingElement", "showBackdrop", "swipeToClose", "translucent", "trigger"]
+  inputs: [
+    'animated',
+    'backdropBreakpoint',
+    'backdropDismiss',
+    'breakpoints',
+    'cssClass',
+    'enterAnimation',
+    'event',
+    'handle',
+    'initialBreakpoint',
+    'isOpen',
+    'keyboardClose',
+    'leaveAnimation',
+    'mode',
+    'presentingElement',
+    'showBackdrop',
+    'swipeToClose',
+    'translucent',
+    'trigger',
+  ],
 })
 export class IonModal {
   @ContentChild(TemplateRef, { static: false }) template: TemplateRef<any>;
@@ -43,6 +91,15 @@ export class IonModal {
       c.detectChanges();
     });
 
-    proxyOutputs(this, this.el, ["ionModalDidPresent", "ionModalWillPresent", "ionModalWillDismiss", "ionModalDidDismiss", "didPresent", "willPresent", "willDismiss", "didDismiss"]);
+    proxyOutputs(this, this.el, [
+      'ionModalDidPresent',
+      'ionModalWillPresent',
+      'ionModalWillDismiss',
+      'ionModalDidDismiss',
+      'didPresent',
+      'willPresent',
+      'willDismiss',
+      'didDismiss',
+    ]);
   }
 }