+++ angular/src/providers/gesture-controller.ts
@@ -1,4 +1,4 @@
-import { Injectable, NgZone } from '@angular/core';
+import { NgZone, Injectable } from '@angular/core';
 import { Gesture, GestureConfig, createGesture } from '@ionic/core/components';
 
 @Injectable({
@@ -11,10 +11,10 @@ export class GestureController {
    */
   create(opts: GestureConfig, runInsideAngularZone = false): Gesture {
     if (runInsideAngularZone) {
-      Object.getOwnPropertyNames(opts).forEach(key => {
-        if (typeof (opts as any)[key] === 'function') {
-          const fn = (opts as any)[key];
-          (opts as any)[key] = (...props: any) => this.zone.run(() => fn(...props));
+      Object.getOwnPropertyNames(opts).forEach((key) => {
+        if (typeof opts[key] === 'function') {
+          const fn = opts[key];
+          opts[key] = (...props: any[]) => this.zone.run(() => fn(...props));
         }
       });
     }