+++ angular/src/providers/popover-controller.ts
@@ -1,4 +1,4 @@
-import { ComponentFactoryResolver, Injectable, Injector } from '@angular/core';
+import { ComponentFactoryResolver, Injector, Injectable } from '@angular/core';
 import { PopoverOptions, popoverController } from '@ionic/core/components';
 
 import { OverlayBaseController } from '../util/overlay';
@@ -7,11 +7,10 @@ import { AngularDelegate } from './angular-delegate';
 
 @Injectable()
 export class PopoverController extends OverlayBaseController<PopoverOptions, HTMLIonPopoverElement> {
-
   constructor(
     private angularDelegate: AngularDelegate,
     private resolver: ComponentFactoryResolver,
-    private injector: Injector,
+    private injector: Injector
   ) {
     super(popoverController);
   }
@@ -19,7 +18,7 @@ export class PopoverController extends OverlayBaseController<PopoverOptions, HTM
   create(opts: PopoverOptions): Promise<HTMLIonPopoverElement> {
     return super.create({
       ...opts,
-      delegate: this.angularDelegate.create(this.resolver, this.injector)
+      delegate: this.angularDelegate.create(this.resolver, this.injector),
     });
   }
 }