+++ angular/src/schematics/add/index.ts
@@ -1,8 +1,21 @@
-import { join, Path } from '@angular-devkit/core';
-import { apply, chain, mergeWith, move, Rule, SchematicContext, SchematicsException, template, Tree, url } from '@angular-devkit/schematics';
+import { Path, join } from '@angular-devkit/core';
+import {
+  Rule,
+  SchematicContext,
+  Tree,
+  apply,
+  chain,
+  mergeWith,
+  move,
+  SchematicsException,
+  template,
+  url,
+} from '@angular-devkit/schematics';
 import { NodePackageInstallTask } from '@angular-devkit/schematics/tasks';
+import { getWorkspace } from '@schematics/angular/utility/workspace';
+
 import { addModuleImportToRootModule } from './../utils/ast';
-import { addArchitectBuilder, addAsset, addStyle, getDefaultAngularAppName, getWorkspace, WorkspaceProject, WorkspaceSchema } from './../utils/config';
+import { addArchitectBuilder, addAsset, addStyle, getDefaultAngularAppName } from './../utils/config';
 import { addPackageToPackageJson } from './../utils/package';
 import { Schema as IonAddOptions } from './schema';
 
@@ -15,24 +28,14 @@ function addIonicAngularToPackageJson(): Rule {
 
 function addIonicAngularToolkitToPackageJson(): Rule {
   return (host: Tree) => {
-    addPackageToPackageJson(
-      host,
-      'devDependencies',
-      '@ionic/angular-toolkit',
-      'latest'
-    );
+    addPackageToPackageJson(host, 'devDependencies', '@ionic/angular-toolkit', 'latest');
     return host;
   };
 }
 
 function addIonicAngularModuleToAppModule(projectSourceRoot: Path): Rule {
   return (host: Tree) => {
-    addModuleImportToRootModule(
-      host,
-      projectSourceRoot,
-      'IonicModule.forRoot()',
-      '@ionic/angular'
-    );
+    addModuleImportToRootModule(host, projectSourceRoot, 'IonicModule.forRoot()', '@ionic/angular');
     return host;
   };
 }
@@ -50,13 +53,13 @@ function addIonicStyles(projectName: string, projectSourceRoot: Path): Rule {
       'node_modules/@ionic/angular/css/text-alignment.css',
       'node_modules/@ionic/angular/css/text-transformation.css',
       'node_modules/@ionic/angular/css/flex-utils.css',
-      `${projectSourceRoot}/theme/variables.css`
-    ]
+      `${projectSourceRoot}/theme/variables.css`,
+    ];
 
-    ionicStyles.forEach(entry => {
+    ionicStyles.forEach((entry) => {
       addStyle(host, projectName, entry);
     });
-   return host;
+    return host;
   };
 }
 
@@ -65,7 +68,7 @@ function addIonicons(projectName: string): Rule {
     const ioniconsGlob = {
       glob: '**/*.svg',
       input: 'node_modules/ionicons/dist/ionicons/svg',
-      output: './svg'
+      output: './svg',
     };
     addAsset(host, projectName, 'build', ioniconsGlob);
     addAsset(host, projectName, 'test', ioniconsGlob);
@@ -79,25 +82,25 @@ function addIonicBuilder(projectName: string): Rule {
       builder: '@ionic/angular-toolkit:cordova-serve',
       options: {
         cordovaBuildTarget: `${projectName}:ionic-cordova-build`,
-        devServerTarget: `${projectName}:serve`
+        devServerTarget: `${projectName}:serve`,
       },
       configurations: {
         production: {
           cordovaBuildTarget: `${projectName}:ionic-cordova-build:production`,
-          devServerTarget: `${projectName}:serve:production`
-        }
-      }
+          devServerTarget: `${projectName}:serve:production`,
+        },
+      },
     });
     addArchitectBuilder(host, projectName, 'ionic-cordova-build', {
       builder: '@ionic/angular-toolkit:cordova-build',
       options: {
-        browserTarget: `${projectName}:build`
+        browserTarget: `${projectName}:build`,
       },
       configurations: {
         production: {
-          browserTarget: `${projectName}:build:production`
-        }
-      }
+          browserTarget: `${projectName}:build:production`,
+        },
+      },
     });
     return host;
   };
@@ -110,22 +113,18 @@ function installNodeDeps() {
 }
 
 export default function ngAdd(options: IonAddOptions): Rule {
-  return (host: Tree) => {
-    const workspace: WorkspaceSchema = getWorkspace(host);
+  return async (host: Tree) => {
+    const workspace = await getWorkspace(host);
     if (!options.project) {
       options.project = getDefaultAngularAppName(workspace);
     }
-    const project: WorkspaceProject = workspace.projects[options.project];
-    if (project.projectType !== 'application') {
-      throw new SchematicsException(
-        `Ionic Add requires a project type of "application".`
-      );
+    const project = workspace.projects.get(options.project);
+
+    if (!project || project.extensions.projectType !== 'application') {
+      throw new SchematicsException(`Ionic Add requires a project type of "application".`);
     }
     const sourcePath: Path = join(project.sourceRoot as Path);
-    const rootTemplateSource = apply(url('./files/root'), [
-      template({ ...options }),
-      move(sourcePath)
-    ]);
+    const rootTemplateSource = apply(url('./files/root'), [template({ ...options }), move(sourcePath)]);
     return chain([
       // @ionic/angular
       addIonicAngularToPackageJson(),
@@ -136,7 +135,7 @@ export default function ngAdd(options: IonAddOptions): Rule {
       addIonicons(options.project),
       mergeWith(rootTemplateSource),
       // install freshly added dependencies
-      installNodeDeps()
+      installNodeDeps(),
     ]);
   };
 }