+++ angular/src/schematics/utils/devkit-utils/change.ts
@@ -10,7 +10,6 @@ export interface Host {
   read(path: string): Promise<string>;
 }
 
-
 export interface Change {
   apply(host: Host): Promise<void>;
 
@@ -26,7 +25,6 @@ export interface Change {
   readonly description: string;
 }
 
-
 /**
  * An operation that does nothing.
  */
@@ -34,15 +32,15 @@ export class NoopChange implements Change {
   description = 'No operation.';
   order = Infinity;
   path = null;
-  apply() { return Promise.resolve(); }
+  apply(): Promise<void> {
+    return Promise.resolve();
+  }
 }
-
 
 /**
  * Will add text to the source code.
  */
 export class InsertChange implements Change {
-
   order: number;
   description: string;
 
@@ -57,8 +55,8 @@ export class InsertChange implements Change {
   /**
    * This method does not insert spaces if there is none in the original string.
    */
-  apply(host: Host) {
-    return host.read(this.path).then(content => {
+  apply(host: Host): Promise<void> {
+    return host.read(this.path).then((content) => {
       const prefix = content.substring(0, this.pos);
       const suffix = content.substring(this.pos);
 
@@ -71,7 +69,6 @@ export class InsertChange implements Change {
  * Will remove text from the source code.
  */
 export class RemoveChange implements Change {
-
   order: number;
   description: string;
 
@@ -84,7 +81,7 @@ export class RemoveChange implements Change {
   }
 
   apply(host: Host): Promise<void> {
-    return host.read(this.path).then(content => {
+    return host.read(this.path).then((content) => {
       const prefix = content.substring(0, this.pos);
       const suffix = content.substring(this.pos + this.toRemove.length);
 
@@ -101,8 +98,7 @@ export class ReplaceChange implements Change {
   order: number;
   description: string;
 
-  constructor(public path: string, private pos: number, private oldText: string,
-              private newText: string) {
+  constructor(public path: string, private pos: number, private oldText: string, private newText: string) {
     if (pos < 0) {
       throw new Error('Negative positions are invalid');
     }
@@ -111,7 +107,7 @@ export class ReplaceChange implements Change {
   }
 
   apply(host: Host): Promise<void> {
-    return host.read(this.path).then(content => {
+    return host.read(this.path).then((content) => {
       const prefix = content.substring(0, this.pos);
       const suffix = content.substring(this.pos + this.oldText.length);
       const text = content.substring(this.pos, this.pos + this.oldText.length);