+++ angular/src/util/ionic-router-reuse-strategy.ts
@@ -1,27 +1,31 @@
 import { ActivatedRouteSnapshot, DetachedRouteHandle, RouteReuseStrategy } from '@angular/router';
 
 export class IonicRouteStrategy implements RouteReuseStrategy {
-
+  // eslint-disable-next-line @typescript-eslint/no-unused-vars
   shouldDetach(_route: ActivatedRouteSnapshot): boolean {
     return false;
   }
 
+  // eslint-disable-next-line @typescript-eslint/no-unused-vars
   shouldAttach(_route: ActivatedRouteSnapshot): boolean {
     return false;
   }
 
-  store(_route: ActivatedRouteSnapshot, _detachedTree: DetachedRouteHandle): void {
+  store(
+    // eslint-disable-next-line @typescript-eslint/no-unused-vars
+    _route: ActivatedRouteSnapshot,
+    // eslint-disable-next-line @typescript-eslint/no-unused-vars
+    _detachedTree: DetachedRouteHandle
+  ): void {
     return;
   }
 
+  // eslint-disable-next-line @typescript-eslint/no-unused-vars
   retrieve(_route: ActivatedRouteSnapshot): DetachedRouteHandle | null {
     return null;
   }
 
-  shouldReuseRoute(
-    future: ActivatedRouteSnapshot,
-    curr: ActivatedRouteSnapshot
-  ): boolean {
+  shouldReuseRoute(future: ActivatedRouteSnapshot, curr: ActivatedRouteSnapshot): boolean {
     if (future.routeConfig !== curr.routeConfig) {
       return false;
     }